refactoring based on PR comments and adding common _types file
