+++ libs/langchain-cloudflare/langchain_cloudflare/rerankers.py
@@ -16,6 +16,8 @@ from langchain_core.documents import Document
 from langchain_core.utils import from_env, secret_from_env
 from pydantic import BaseModel, ConfigDict, Field, PrivateAttr, SecretStr
 
+from langchain_cloudflare._errors import TokenErrors
+
 # MARK: - Constants
 DEFAULT_RERANKER_MODEL = "@cf/baai/bge-reranker-base"
 
@@ -121,18 +123,10 @@ class CloudflareWorkersAIReranker(BaseModel):
 
         # Validate credentials
         if not self.account_id:
-            raise ValueError(
-                "A Cloudflare account ID must be provided either through "
-                "the account_id parameter or CF_ACCOUNT_ID environment variable. "
-                "Or pass the 'binding' parameter (env.AI) in a Python Worker."
-            )
+            raise ValueError(TokenErrors.NO_ACCOUNT_ID_SET)
 
-        if not any([self.api_token, self.api_token.get_secret_value()]):
-            raise ValueError(
-                "A Cloudflare API token must be provided either through "
-                "the api_token parameter or CF_AI_API_TOKEN environment variable. "
-                "Or pass the 'binding' parameter (env.AI) in a Python Worker."
-            )
+        if not self.api_token or not self.api_token.get_secret_value():
+            raise ValueError(TokenErrors.INSUFFICIENT_AI_TOKENS)
 
         self.headers = {"Authorization": f"Bearer {self.api_token.get_secret_value()}"}
